requirements.txt
Django==3.0.7
django-cleanup==5.0.0
django-crispy-forms==1.9.1
Pillow==7.1.2
psycopg2==2.8.5
requests==2.24.0
requests-oauthlib==1.3.0
storage==0.0.4.3
